+++ launch.py
@@ -165,7 +165,10 @@ if __name__ == "__main__":
     installer.args = args
     installer.setup_logging()
     installer.log.info('Starting SD.Next')
-    sys.excepthook = installer.custom_excepthook
+    try:
+        sys.excepthook = installer.custom_excepthook
+    except Exception:
+        pass
     installer.read_options()
     if args.skip_all:
         args.quick = True